feat: 사전/사후평가 응답 DB에 저장, 사전 평가 기록이 존재할 경우 예외 처리

+++ src/main/java/com/education/takeit/exam/repository/ExamRepository.java
@@ -3,6 +3,6 @@ package com.education.takeit.exam.repository;
 import com.education.takeit.exam.entity.Exam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExamRepository extends JpaRepository<Exam, Integer> {
+public interface ExamRepository extends JpaRepository<Exam, Long> {
   Exam findByExamId(Long examId);
 }
